CA-223797: Callback when watch-queue overflows
A user of the unix client module can register a callback function
which will be called on overflow of the queue of watch events (the
ones which are waiting to be handled by the watch callback).

No other change at this stage: the queue cannot recover after the
overflow and the only remedy (to start receiving watch events again)
would be to discard the Client instance and start afresh with a new
one.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>